Weapon store fully functional
Weapon concealing action makes a complete copy of unit's weapon,
including ammo and accessories, which can be accessed easily again via
another action. Prioritises rifle first but will also get pistol out if
that is the only item. Action works on AI and player. Detects if storing
is possible and does not appear if not.